abstract to connectionitem component
