+++ src/components/Dashboard/ConnectionsList.tsx
@@ -1,12 +1,9 @@
-import { Avatar, Card, Col, Row, Space, Typography } from "antd";
+import { Card, Row } from "antd";
 import React, { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { useSelector } from "react-redux";
 import { Call } from "src/types/Call";
 import { Connection } from "src/types/Connection";
-import { genFullName } from "src/utils/utils";
-import { selectEndedCalls } from "src/redux/selectors";
-import { differenceInDays } from "date-fns";
+import ConnectionItem from "./ConnectionItem";
 
 interface Props {
   calls: Call[];
@@ -15,8 +12,6 @@ interface Props {
 
 const ConnectionsList: React.FC<Props> = ({ calls, connections }: Props) => {
   const { t } = useTranslation("dashboard");
-  const endedCalls = useSelector(selectEndedCalls);
-
   const tabList = [
     {
       key: "approved",
@@ -30,20 +25,6 @@ const ConnectionsList: React.FC<Props> = ({ calls, connections }: Props) => {
 
   const [activeContactTab, setActiveContactTab] = useState("approved");
 
-  // TODO: Remove this messy function and pass the days past in the api endpoint
-  const getDaysPastNum = (connectionId: number) => {
-    if (!endedCalls) return 0;
-    const last = Math.max.apply(
-      Math,
-      endedCalls
-        .filter((call) => call.connectionId === connectionId)
-        .map((call) => call.end)
-    );
-    const lastCall = endedCalls.find((call) => call.end === last);
-    if (lastCall) return differenceInDays(new Date(lastCall.end), new Date());
-    return 0;
-  };
-
   return (
     <Card
       title={t("connection.title")}
@@ -55,34 +36,7 @@ const ConnectionsList: React.FC<Props> = ({ calls, connections }: Props) => {
         {connections
           .filter((connection) => connection.status === activeContactTab)
           .map((connection) => (
-            <Col
-              key={connection.id}
-              className="d-flex flex-column align-items-center"
-            >
-              <Space direction="vertical">
-                <Avatar
-                  shape="square"
-                  size={80}
-                  src={connection.user.profileImgPath}
-                />
-                <div>
-                  <div>
-                    <Typography.Text>
-                      {genFullName(connection.user)}
-                    </Typography.Text>
-                  </div>
-                  {connection.status === "approved" && (
-                    <div>
-                      <Typography.Text type="secondary">
-                        {t("connection.lastCall", {
-                          daysPastNum: getDaysPastNum(connection.id),
-                        })}
-                      </Typography.Text>
-                    </div>
-                  )}
-                </div>
-              </Space>
-            </Col>
+            <ConnectionItem connection={connection} />
           ))}
       </Row>
     </Card>
