+++ src/i18n/es/modals.json
@@ -1,7 +1,7 @@
 {
   "cancelCallModal": {
-    "title": "Cancel Call?",
-    "okText": "Confirm",
+    "title": "Cancelar llamada?",
+    "okText": "Confirmar",
     "messageOne": "You are about to cancel your call with {{connectionName}} for:",
     "messageTwo": "Please select a cancellation reason to send to {{connectionFirstName}}.",
     "dropdownPlaceholder": "Cancellation reason"
